Wed, Nov 8, 2023, 8:33 PM -08:00

+++ 2324/chapter2/examples/calc/calc.c
@@ -7,7 +7,7 @@
 // mod opperator --> Usage: ./calc [num1] [operator] [num2]
 // index:                      0      1        2       3
 
-float mod(float num1, float num2);
+double mod(double num1, double num2);
 
 int main(int argc, string argv[])
 {
@@ -75,7 +75,7 @@ int main(int argc, string argv[])
     }
 }
 
-double mod(float num1, float num2)
+double mod(double num1, double  num2)
 {
     if (num2 == 0)
     {
